Add ElmoMasterNode.c and .h
We want them in the repository so that we don't have to rely on
having CanFestival's objdictgen to remake them. This is important
on modern systems which do not provide Python 2, since objdictgen
*does not work* with later versions of Python. However, this
shouldn't be much of a sacrifice because it's unlikely we will
ever need to change ElmoMasterNode.od, and hence will not need to
run objdictgen again. Hopefully.