Move ADC trigger arming into startRadar()
We now delay DMA thread start and trigger arming for the ADCs until
we're in startRadar() and have put ourselves at a time where we know
with extremely high certainty the correct time of the next 1PPS signal
that will arrive and start the ADCs, and hence we will have a correct
radar start time.

Previously, it was assumed that the ADCs would start on the first 1PPS
after the call to NAV_GlobalGateOpen(). In fact, the ADCs start on the
first 1PPS after the call to NAV_TrigArm(). Moving the NAV_TrigArm()
call into startRadar() means that we now always start on the intended
1PPS signal.